Using fp16.

+++ Smoke3D/Content/CSAdvect3D.hlsl
@@ -7,16 +7,16 @@
 //--------------------------------------------------------------------------------------
 // Constants
 //--------------------------------------------------------------------------------------
-static const float2 g_vDecay = { 0.996, 1.0 };
+static const min16float2 g_vDecay = { 0.996, 1.0 };
 
 //--------------------------------------------------------------------------------------
 // Textures
 //--------------------------------------------------------------------------------------
-RWTexture3D<half3>	g_rwPhiVel		: register (u0);
-RWTexture3D<half>	g_rwPhiDen		: register (u1);
-Texture3D<half3>	g_roPhiVel		: register (t0);
-Texture3D<half>		g_roPhiDen		: register (t1);
-Texture3D<half3>	g_roVelocity	: register (t2);
+RWTexture3D<min16float3>	g_rwPhiVel		: register (u0);
+RWTexture3D<min16float>		g_rwPhiDen		: register (u1);
+Texture3D<min16float3>		g_roPhiVel		: register (t0);
+Texture3D<min16float>		g_roPhiDen		: register (t1);
+Texture3D<min16float3>		g_roVelocity	: register (t2);
 
 //--------------------------------------------------------------------------------------
 // Texture samplers
@@ -31,8 +31,8 @@ SamplerState		g_smpLinear		: register (s1);
 void main(uint3 DTid : SV_DispatchThreadID)
 {
 	// Velocity tracing
-	const half3 vU = g_roVelocity[DTid];
-	const half3 vTex = (DTid + 0.5) * g_vTexel - vU * g_fDeltaTime;
+	const float3 vU = g_roVelocity[DTid];
+	const float3 vTex = (DTid + 0.5) * g_vTexel - vU * g_fDeltaTime;
 
 	// Update velocity and density
 	g_rwPhiVel[DTid] = g_roPhiVel.SampleLevel(g_smpLinear, vTex, 0);
